requirements.txt
biopython
torch
numpy
fair-esm